% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOC.R
\name{AOC}
\alias{AOC}
\title{Assessing Order Completion}
\usage{
AOC(Salesfile, AssessingFile, Output = FALSE)
}
\arguments{
\item{Salesfile}{the datapath of the new sales file of Bonny&Read}

\item{AssessingFile}{the datapath of the coworked replenishment schedule}

\item{Output}{Output the result to desktop as csv file}
}
\description{
This function aims to address the issue that stocking departmnet
do not perceive of whether the order complete or not. This can output the
orderingid which has been completely ready for sent out.
}

