% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA.R
\name{MPA}
\alias{MPA}
\title{Monthly Performance Analysis}
\usage{
MPA(Dataroute, Year, Month)
}
\arguments{
\item{Dataroute}{this should be the datapath of the folder which conprise
all older format sales files.}

\item{Year}{Assign the year you want to check for MPA, the format are supposed
to be YYYY format}

\item{Month}{Assign the month you want to check for MPA, the format are supposed
to be a a range from 1 to 12.}
}
\description{
This function allows you to report some specific indicators of
Bonny&Read corp. performance, like revenue,retention rate,etc.
}

