% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRS.R
\name{RRS}
\alias{RRS}
\title{Replenishment Recommendation System}
\usage{
RRS(AssignedPath, StockingData, pagelimit)
}
\arguments{
\item{AssignedPath}{The folder path which stores salesfiles}

\item{pagelimit}{you should assign the total page number of the product pages
of the offcial website.}

\item{Stockingdata}{The path of the latest storefiles}
}
\description{
This application is used to analyze sale force of each items
and establish recommendation level which base on own algorithm
}
\author{
Avalon Lin
}

